+++ src/main/kotlin/com/github/hovrawl/jetbrainstoggleabletoolwindows/actions/ToggleBottomStripeToolWindow.kt
@@ -2,4 +2,4 @@ package com.github.hovrawl.jetbrainstoggleabletoolwindows.actions
 
 import com.intellij.openapi.wm.ToolWindowAnchor
 
-class ToggleBottomStripeToolWindow : ToggleStripeAction(ToolWindowAnchor.BOTTOM)
+class ToggleBottomIslandToolWindow : ToggleIslandAction(ToolWindowAnchor.BOTTOM)