+++ src/main/kotlin/com/github/hovrawl/jetbrainstoggleabletoolwindows/actions/ToggleLeftStripeToolWindow.kt
@@ -2,4 +2,4 @@ package com.github.hovrawl.jetbrainstoggleabletoolwindows.actions
 
 import com.intellij.openapi.wm.ToolWindowAnchor
 
-class ToggleLeftStripeToolWindow : ToggleStripeAction(ToolWindowAnchor.LEFT)
+class ToggleLeftIslandToolWindow : ToggleIslandAction(ToolWindowAnchor.LEFT)