+++ src/main/kotlin/com/github/hovrawl/jetbrainstoggleabletoolwindows/actions/ToggleRightStripeToolWindow.kt
@@ -2,4 +2,4 @@ package com.github.hovrawl.jetbrainstoggleabletoolwindows.actions
 
 import com.intellij.openapi.wm.ToolWindowAnchor
 
-class ToggleRightStripeToolWindow : ToggleStripeAction(ToolWindowAnchor.RIGHT)
+class ToggleRightIslandToolWindow : ToggleIslandAction(ToolWindowAnchor.RIGHT)